Ids Conv: Casts values to float before operation in generic_scale()
Former-commit-id: af08d3782afef20c5ea17953effdbecde37dd75b

+++ webapp/ids/ids_converter.py
@@ -493,7 +493,7 @@ class IdsConverter(object):
 		"""
 
 		# pylint: disable-msg=C0103; (Snake-case naming)
-		X = numpy.array(values)
+		X = numpy.array(values, dtype=numpy.float_)
 
 		if X.ndim != 2:
 			raise ValueError("generic_scale() requires a two-dimensional input")
